+++ R/growthModels.R
@@ -23,31 +23,33 @@
 growthModels <- function(...) {
   op <- par(mar=c(0,0,2,0))
   plot(1,type="n",ylim=c(0,10),xlim=c(0,1),xaxt="n",yaxt="n",xlab="",ylab="",bty="n",main="FSA Growth Model Parametrizations",...)
-  iGrowthModels("vbOriginal",0,9.5)
-  iGrowthModels("vbTypical",0,8)
-  iGrowthModels("vbGallucciQuinn",0,6.5)
-  iGrowthModels("vbMooij",0,5)
-  iGrowthModels("vbSchnute",0,3.5)
-  iGrowthModels("vbFrancis",0,1.75)
-  iGrowthModels("vbFrancis2",0.15,0.5)
+  iGrowthModels("vbOriginal",     0,9.0)
+  iGrowthModels("vbTypical",      0,7.5)
+  iGrowthModels("vbGallucciQuinn",0,6.0)
+  iGrowthModels("vbMooij",        0,4.5)
+  iGrowthModels("vbWeisberg",     0,3.0)
   abline(v=0.55)
-  iGrowthModels("Gompertz1",0.55,9.5)
-  iGrowthModels("Gompertz2",0.55,8)
-  iGrowthModels("Gompertz3",0.55,6.5)
+  iGrowthModels("vbSchnute", 0.55,9.0)
+  iGrowthModels("vbFrancis", 0.55,7.25)
+  iGrowthModels("vbFrancis2",0.70,6.0)
+  iGrowthModels("Gompertz1", 0.55,3.5)
+  iGrowthModels("Gompertz2", 0.55,2.0)
+  iGrowthModels("Gompertz3", 0.55,0.5)
   par(op)
 }
 
 vbModels <- function(...) {
   op <- par(mar=c(0,0,3,0))
   plot(1,type="n",ylim=c(0,7),xlim=c(0,1),xaxt="n",yaxt="n",xlab="",ylab="",bty="n",main="FSA von Bertalanffy Parametrizations",...)
-  iGrowthModels("vbOriginal",0,6.5)
-  iGrowthModels("vbTypical",0,4.5)
-  iGrowthModels("vbGallucciQuinn",0,2.5)
+  iGrowthModels("vbOriginal",     0,6.0)
+  iGrowthModels("vbTypical",      0,4.0)
+  iGrowthModels("vbGallucciQuinn",0,2.0)
+  iGrowthModels("vbMooij",        0,0.5)
   abline(v=0.5)
-  iGrowthModels("vbMooij",0.5,6.5)
-  iGrowthModels("vbSchnute",0.5,4.5)
-  iGrowthModels("vbFrancis",0.5,2.5)
-  iGrowthModels("vbFrancis2",0.65,1)
+  iGrowthModels("vbWeisberg",0.50,6.0)
+  iGrowthModels("vbSchnute", 0.50,4.0)
+  iGrowthModels("vbFrancis", 0.50,2.0)
+  iGrowthModels("vbFrancis2",0.65,0.5)
   par(op)
 }
 
@@ -58,6 +60,7 @@ iGrowthModels <- function(which,xpos,ypos) {
    vbTypical=       {text(xpos,ypos,expression(plain("vbTypical: ")~~~E(L[t])==L[infinity]*bgroup("(",1-e^{-K*(t~-~t[0])},")")),pos=4)},
    vbGallucciQuinn= {text(xpos,ypos,expression(plain("vbGallucciQuinn: ")~~~E(L[t])==frac(omega,K)*~bgroup("(",1-e^{-K*(t~-~t[0])},")")),pos=4)},
    vbMooij=         {text(xpos,ypos,expression(plain("vbMooij: ")~~~E(L[t])==L[infinity]~-~(L[infinity]-L[0])*~e^{-frac(omega,L[infinity])*~t}),pos=4)},
+   vbWeisberg=      {text(xpos,ypos,expression(plain("vbWeisberg: ")~~~E(L[t])==L[infinity]*bgroup("(",1-e^{-frac(log2,K[0])*(t~-~t[0])},")")),pos=4)},
    vbSchnute=       {text(xpos,ypos,expression(plain("vbSchnute: ")~~~E(L[t])==L[1]+(L[3]-L[1])*~frac(1-e^{-K*(~t~-~t[1])},1-e^{-K*(~t[3]~-~t[1])})),pos=4)},
    vbFrancis=       {text(xpos,ypos,expression(plain("vbFrancis: ")~~~E(L[t])==L[1]+(L[3]-L[1])*~frac(1-r^{2*frac(t-t[1],t[3]-t[1])},1-r^{2})),pos=4)},
    vbFrancis2=      {text(xpos,ypos,expression(plain("where" )~r==frac(L[3]-L[2],L[2]-L[1])),pos=4)},